requirements.txt
fastapi
uvicorn
python-dotenv
requests
selenium
webdriver-manager
beautifulsoup4
slowapi
bcrypt<4.0.0
passlib[bcrypt]
